// apps/yaniv/src/engine/types.ts
export type HandCard =
  | { kind: 'card'; suit: Suit; rank: Rank }
  | { kind: 'joker'; jid: number }